add get module detail service, refactor module card into its own file
