+++ mobile/services/courses.ts
@@ -1,5 +1,5 @@
 import apiClient from './api';
-import { Course, CourseDetail, Module, Post } from '../types';
+import { Course, CourseDetail, Post } from '../types';
 
 export async function getUserCourses(): Promise<Course[]> {
     // In starter-code, course-group relationships were removed
@@ -16,29 +16,7 @@ export async function getCourseDetail(courseId: number): Promise<CourseDetail> {
     return response.data;
 }
 
-export async function getModuleDetail(moduleId: number): Promise<Module> {
-    try {
-        if (__DEV__) {
-            console.log(`Fetching module detail for module ID: ${moduleId}`);
-        }
-        const response = await apiClient.get<Module>(
-            `/api/modules/${moduleId}`
-        );
-        if (__DEV__) {
-            console.log('Module detail response:', response.data);
-        }
-        return response.data;
-    } catch (error) {
-        if (__DEV__) {
-            if (error instanceof Error) {
-                console.error('Error fetching module detail:', error.message);
-            } else {
-                console.error('Error fetching module detail:', error);
-            }
-        }
-        throw error;
-    }
-}
+export { getModuleDetail } from './modules';
 
 export async function getModulePosts(moduleId: number): Promise<Post[]> {
     try {
